Refactor the testcases
